+++ src/app/core/interceptors/unauthorized-error.interceptor.ts
@@ -1,7 +1,7 @@
 import {HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
 import {Inject, Injectable} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
-import {Environment, EnvironmentToken} from '../environment';
+import {Environment, EnvironmentToken} from '../../environment';
 import 'rxjs/add/operator/catch';
 
 @Injectable()
